+++ ignite/pkg/chainregistry/doc.go
@@ -1,0 +1,3 @@
+// package chainregistry is a package that contains the chain.json and assetlist.json structs from the chain registry.
+// Useful when parsing or creating chain.json and assetlist.json files.
+package chainregistry
